[DDW-651]: Update dropdown menu component on delegation center screen
