+++ source/renderer/app/components/staking/delegation-center/DropdownMenu.scss
@@ -1,0 +1,91 @@
+.component {
+  :global {
+    .SimpleSelect_select {
+      .SimpleSelect_selectInput {
+        &::after {
+          display: none;
+        }
+        input {
+          border: 0;
+          padding: 0;
+          width: 22px;
+          height: 22px;
+        }
+        .SimpleInput_customValueBlock {
+          height: auto;
+        }
+      }
+    }
+
+    .SimpleOptions_options {
+      left: auto !important;
+      position: absolute;
+      right: 0;
+      top: 25px !important;
+      width: 200px !important;
+
+      ul {
+        li {
+          background: var(--theme-staking-dropdown-item-background-color);
+          border-radius: 4px;
+          color: var(--theme-staking-font-color-regular);
+          padding: 6px 9px 6px 25px;
+          text-align: right;
+
+          &.highlightedOption {
+            background-color: var(
+              --theme-staking-dropdown-item-background-color
+            );
+          }
+
+          &::after {
+            display: none;
+          }
+
+          &:hover {
+            background-color: var(
+              --theme-staking-dropdown-item-background-color-hover
+            );
+            span {
+              color: var(--theme-staking-dropdown-item-text-color-hover);
+            }
+          }
+
+          span {
+            white-space: nowrap;
+          }
+
+          &:first-child {
+            border-radius: 4px 4px 0 0;
+          }
+
+          &:last-child {
+            border-radius: 0 0 4px 4px;
+          }
+        }
+        .SimpleOptions_option.SimpleOptions_selectedOption {
+          display: list-item;
+        }
+      }
+    }
+
+    .SimpleBubble_root:not(.SimpleBubble_transparent) {
+      .SimpleBubble_bubble {
+        border: none;
+        border-radius: 3px;
+        box-shadow: 0 5px 20px 0 rgba(0, 0, 0, 0.12);
+        min-width: 100px;
+        right: 15px;
+      }
+      [data-bubble-arrow]:before,
+      [data-bubble-arrow]:after {
+        content: '';
+        display: none;
+      }
+    }
+  }
+
+  .dropdownToggle {
+    margin-left: 10px;
+  }
+}
