--- source/renderer/app/components/staking/delegation-center/DropdownMenu/DropdownToggle.scss
@@ -1,14 +1,0 @@
-@import '../../stakingConfig';
-
-.component {
-  width: 100%;
-
-  .label {
-    @extend %regularText;
-
-    align-items: center;
-    display: flex;
-    justify-content: flex-end;
-    padding-bottom: 3px;
-  }
-}
